Add origin rebasing for orbital coordinates
Introduce WorldPosition for double-precision kilometer coordinates and OriginAnchor for selecting the entity that drives origin rebasing.

Add OriginRebaseSystem to keep Transform positions in local meter space, rebase the origin in 3 km steps around the anchor, and expose helpers for local/world conversion.

Move OrbitalEngine to WorldPosition, double GM values, and kilometer-per-second orbital velocities while converting orbit preview lines back to local meters for rendering.

Wire the demo satellite as the origin anchor, initialize celestial/satellite world positions, keep physics bodies synchronized after rebases, and expose world position/orbital double values in ImGui.

Document the large-world coordinate model and add doctest coverage for world-to-local conversion and anchor-driven rebasing.

+++ sources/components/Celestial.hpp
@@ -3,6 +3,6 @@
 #include <cmath>
 
 struct Celestial {
-    float GM = 1.0f;
-    float radius = INFINITY;
+    double GM = 1.0;
+    double radius = INFINITY;
 };
